fix(coinley): address security, localization, and code quality issues
Security:
- Validate paymentId matches stored payment to prevent arbitrary booking marking

Bug fixes:
- Fix chargeCard using undefined paymentId (now uses externalId)
- Fix requirePayment state not updating in components (read from getAppData)
- Allow decimal prices in zod validation

Localization:
- Use t() for all user-facing strings in components

Code quality:
- Add Prisma select to handlePaymentRefunded for efficiency
- Replace any types with proper type definitions in webhook handlers

+++ packages/app-store/coinley/zod.ts
@@ -34,7 +34,7 @@ export type AppKeysSchema = z.infer<typeof appKeysSchema>;
 export const appDataSchema = eventTypeAppCardZod.merge(
   z.object({
     // Payment settings
-    price: z.number().int().positive({ message: "Price must be greater than 0" }),
+    price: z.number().positive({ message: "Price must be greater than 0" }),
     currency: z.enum(SUPPORTED_CURRENCIES).default("USDT"),
 
     // Network selection
